Update Challenge-5.txt
foldl function, also called the left fold. It folds the list up from the left side. The binary function is applied between the starting value and the head of the list. That produces a new accumulator value and the binary function is called with that value and the next element, etc. 

I use this combine with lcm(Least common multiple)